+++ docs/examples/turboencabulator/src/lib.rs
@@ -1,2 +1,6 @@
-mod generated;
-pub use generated::*;
+//! Turbo Encabulator Registers
+include!(concat!(env!("OUT_DIR"), "/turbo_encab/mod.rs"));
+
+fn main() {
+    TurboEncab;
+}